+++ lib/on_text_input_widget.dart
@@ -1,3 +1,12 @@
 library on_text_input_widget;
 
-export 'src/on_text_input_widget.dart';
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+
+part 'src/on_text_input_widget.dart';
+
+part 'src/widgets/animated_size.dart';
+part 'src/widgets/svg.dart';
+part 'src/widgets/user_field.dart';