+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 ## 0.0.1
 
-* on_text_input_widget with many functionalities
+- on_text_input_widget with many functionalities
 
 ## 0.0.2
 
-* Some size error fixed
+- Some size error fixed
 
 ## 0.0.3
 
-* Some size error fixed
+- Some size error fixed
+
+## 0.0.4
+
+- OnTextInputWidgetUserField added. It is a sub-widget implemented with OnProcessButtonWidget